backend/app/vector_search.py
from google.cloud import aiplatform
import os
from typing import List

class AetherVectorSearch:
    def __init__(self):
        self.project_id = os.getenv("GOOGLE_CLOUD_PROJECT")
        self.location = os.getenv("GOOGLE_CLOUD_LOCATION", "us-central1")
        self.index_id = os.getenv("VERTEX_AI_INDEX_ID")
        self.endpoint_id = os.getenv("VERTEX_AI_ENDPOINT_ID")
        
        if self.project_id:
            aiplatform.init(project=self.project_id, location=self.location)

    def search_similar_scents(self, embedding: List[float], num_neighbors: int = 5):
        """
        Search for similar fragrances using Vertex AI Vector Search.
        """
        if not self.endpoint_id:
            return []
            
        index_endpoint = aiplatform.MatchingEngineIndexEndpoint(
            index_endpoint_name=self.endpoint_id
        )
        
        responses = index_endpoint.find_neighbors(
            deployed_index_id="aether_scent_index",
            queries=[embedding],
            num_neighbors=num_neighbors
        )
        
        return responses

    def get_embedding(self, text: str) -> List[float]:
        """
        Placeholder for text embedding generation.
        In a real scenario, use Vertex AI Text Embeddings API.
        """
        # This would call aiplatform.TextEmbeddingModel
        return [0.1] * 768 # Mock embedding
