backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-aiplatform
supabase
python-dotenv
pytest
httpx
